Added Camera
Camera lerps to the player's position on an infinite place.  All the
other objects render correctly based on camera's position, and the mouse
functionality works correctly.

+++ Game/Hidden_Armada/Hidden_Armada/source/game/objects/Asteroids.cpp
@@ -16,7 +16,7 @@ void Asteroids::Initialize( void )
 	this->SetImgSource(source);
 	this->SetSize(30);
 	this->SetImgCenter(D3DXVECTOR2(32,32));
-	this->SetDir(D3DXVECTOR2(rand()%200 - 100, rand()%200 - 100));
+	this->SetDir(D3DXVECTOR2((float)(rand()%200 - 100), (float)(rand()%200 - 100)));
 
 	if(rand()%2 == 0)
 		m_RotDir = -1.0f;
